Moved ring-based service discovery to pkg/util/servicediscovery/ and removed redundant notifications interfaces
